9.25.17 Accomplishments

+++ CodeEnv.Master.GameContent/Game/Framework/ShipHullEquipment.cs
@@ -37,7 +37,7 @@ namespace CodeEnv.Master.GameContent {
 
         public float Science { get { return Stat.Science; } }
         public float Culture { get { return Stat.Culture; } }
-        public float Income { get { return Stat.Income; } }
+        public decimal Income { get { return Stat.Income; } }
 
         protected new ShipHullStat Stat { get { return base.Stat as ShipHullStat; } }
